Move heroes update into download activity
Fasten collection loading
