+++ app/src/main/java/c4stor/com/feheroes/activities/download/DownloadDataActivity.java
@@ -13,6 +13,7 @@ import android.widget.Toast;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
@@ -20,17 +21,27 @@ import java.net.URL;
 import java.util.Locale;
 
 import c4stor.com.feheroes.R;
+import c4stor.com.feheroes.activities.ModelSingleton;
 import c4stor.com.feheroes.activities.ivcheck.IVCheckActivity;
-
-import static java.lang.Thread.sleep;
+import c4stor.com.feheroes.model.hero.HeroCollection;
+import c4stor.com.feheroes.model.hero.HeroInfo;
+import c4stor.com.feheroes.model.hero.HeroRoll;
 
 public class DownloadDataActivity extends AppCompatActivity {
 
+
+    private ModelSingleton singleton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_download_data);
+        try {
+            singleton = ModelSingleton.getInstance(this);
+        } catch (IOException e) {
+            this.finish();
+        }
+        updateHeroAttributes();
 
         final DownloadTask heroDownloadTask = new DownloadTask(this, "hero.data", false);
         heroDownloadTask.execute("https://nicolasdalsass.github.io/heroesjson/v170630");
@@ -134,5 +145,26 @@ public class DownloadDataActivity extends AppCompatActivity {
                 startActivity(ivCheckIntent);
             }
         }
+    }
+
+    //this method is there to update old HeroCollections
+    private void updateHeroAttributes() {
+        for (HeroRoll heroRoll : singleton.collection) {
+            if (heroRoll.hero.movementType == null) {
+                HeroInfo mapHero = singleton.heroMap.get(heroRoll.hero.name);
+                heroRoll.hero.movementType = mapHero.movementType;
+                heroRoll.hero.weaponType = mapHero.weaponType;
+            }
+            if (heroRoll.hero.atkGrowth == 0) {
+                HeroInfo mapHero = singleton.heroMap.get(heroRoll.hero.name);
+                heroRoll.hero.hpGrowth = mapHero.hpGrowth;
+                heroRoll.hero.atkGrowth = mapHero.atkGrowth;
+                heroRoll.hero.spdGrowth = mapHero.spdGrowth;
+                heroRoll.hero.defGrowth = mapHero.defGrowth;
+                heroRoll.hero.resGrowth = mapHero.resGrowth;
+                heroRoll.hero.availability = mapHero.availability;
+            }
+        }
+        singleton.collection.save(getBaseContext());
     }
 }
